2002-09-16  Martin Baulig  <martin@gnome.org>
	* gnome/Modules.cs (Module): Make this class public.

	* sample/CanvasExample.cs: Insert a missing `break' in a switch
	section to make this compile.

	* sample/Scribble.cs: Added missing casts from EventMask to int.

svn path=/trunk/gtk-sharp/; revision=7493

+++ sample/Scribble.cs
@@ -30,10 +30,10 @@ namespace GtkSamples {
 			darea.MotionNotifyEvent += new MotionNotifyEventHandler (MotionNotifyEvent);
 			darea.ButtonPressEvent += new ButtonPressEventHandler (ButtonPressEvent);
 			darea.Events = (int)EventMask.ExposureMask |
-							   EventMask.LeaveNotifyMask |
-							   EventMask.ButtonPressMask |
-							   EventMask.PointerMotionMask |
-							   EventMask.PointerMotionHintMask;
+							   (int)EventMask.LeaveNotifyMask |
+							   (int)EventMask.ButtonPressMask |
+							   (int)EventMask.PointerMotionMask |
+							   (int)EventMask.PointerMotionHintMask;
 
 			win.ShowAll ();
 			Application.Run ();
